requirements.txt
psd-tools2==1.7.24
Pillow==4.1.1